# openapi.yaml
openapi: 3.0.0
info:
  title: CRUD Backend API
  version: 1.0.0
  description: API para operações CRUD com separação clara de camadas
servers:
  - url: http://localhost:3000/api
    description: Servidor de desenvolvimento
paths:
  /users:
    get:
      summary: Listar todos os usuários
      operationId: listUsers
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Erro interno do servidor
    post:
      summary: Criar um novo usuário
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos fornecidos
        '500':
          description: Erro interno do servidor

  /users/{id}:
    get:
      summary: Obter um usuário por ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado
        '500':
          description: Erro interno do servidor
    put:
      summary: Atualizar um usuário
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado
        '400':
          description: Dados inválidos fornecidos
        '500':
          description: Erro interno do servidor
    delete:
      summary: Excluir um usuário
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Usuário excluído com sucesso
        '404':
          description: Usuário não encontrado
        '500':
          description: Erro interno do servidor

components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email